backend/pyproject.toml
[project]
name = "ads-quality-rater"
version = "1.0.0"
description = "KI-basierter Quality Rater mit Crew AI + Brand Agent"
authors = [{name = "flin Team", email = "team@flin.com"}]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "crewai>=0.70.0",
    "crewai-tools>=0.12.0",
    "google-generativeai>=0.8.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "playwright>=1.48.0",
    "trafilatura>=1.12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
